requirements.txt
diffusers
transformers
accelerate
torch
opencv-python
huggingface_hub
Pillow==9.5.0